backend-flask/requirements.txt
flask
flask-cors

python-dotenv
opentelemetry-instrumentation
opentelemetry-distro 
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

aws-xray-sdk

watchtower



